+++ ProjectKTLT/Student.h
@@ -38,7 +38,7 @@ struct student
     void loadStudentProfile(std::string username); // done
     std::string viewStudentProfile(); // done
     std::string** viewAllCoursesInformations(Static* a, int &n); // done
-    float calculateOverall(float final, float midterm, float other);//support function
+    double calculateOverall(double final, double midterm, double other);//support function
     std::string** getStudentScoreBoard(Static* a, int& n); // done
     void dellocate2DArray(std::string** arr, int n);
